refactor: memindahkan folder components ke folder luar
